requirements.txt
PyYAML>=6.0
pydantic>=2.5.0